+++ astropy/io/fits/header.py
@@ -1094,6 +1094,9 @@ class Header(object):
 
         if other is None:
             pass
+        elif isinstance(other, Header):
+            for card in other.cards:
+                self._update(card)
         elif hasattr(other, 'items'):
             for k, v in iteritems(other):
                 update_from_dict(k, v)